chore(config): restore full GPU pipeline parameters
Revert CPU-era compromises: full feature set, lookback 20, 200 trajectory windows, DT d_model=192/4 layers/6 heads, 50 epochs, 5 seeds, float32 trajectory storage, and full ablation suite with K/RTG retraining.

+++ src/trajectories.py
@@ -143,8 +143,8 @@ def generate_trajectories(cfg: Config) -> Path:
     out_path = output_dir / "trajectories.npz"
     np.savez_compressed(
         out_path,
-        states=states_arr.astype(np.float16),
-        actions=actions_arr.astype(np.float16),
+        states=states_arr,
+        actions=actions_arr,
         rewards=rewards_arr.astype(np.float32),
         rtg=rtg_arr.astype(np.float32),
         lengths=lengths,
@@ -178,8 +178,8 @@ class TrajectoryDataset(Dataset):
         normalize_rtg: bool = True,
     ):
         data = np.load(traj_path)
-        self.states = data["states"].astype(np.float32)
-        self.actions = data["actions"].astype(np.float32)
+        self.states = data["states"]
+        self.actions = data["actions"]
         self.rewards = data["rewards"]
         self.rtg = data["rtg"]
         self.lengths = data["lengths"]
